Added 2 cursor images
Added Client and NetworkRead files for the server interaction. Currently rather basic, just sends and recieves strings.

Added some comments to Game.

Added mousemotion listeners to Scene. Will be used for the cursor images. Got rid of useless imports.

Added enemyInRange function to tell if an enemy is within attack range or not. Also added a cursor function, currently does nothing but it will change cursor according to where it is.

Added comment to Tile

Added comment to Unit

Changed a bunch of variables to private and added some warning suppressors.